+++ FirebaseCore/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Firebase 6.33.0
+- [fixed] Swift Package Manager - Define system framework and system library dependencies. This
+  resolves undefined symbol issues for system dependencies. (#6408, #6413)
+
 # Firebase 6.32.0
 - [changed] Swift Package Manager - It's no longer necessary to select the Firebase or
   FirebaseCore products. Their build targets are implicitly selected when choosing any other